Split ViewHierarchyNode to separate file
